feat: 코드 구조 개선
- index.ts에 구현된 코드들을 별도 파일로 대체. index.ts는 외부 export 용도로만 활용
